[backend] in-pir turned to internal relationship (#11149)
