+++ opencti-platform/opencti-graphql/src/modules/attributes/internalRelationship-registrationAttributes.ts
@@ -1,7 +1,7 @@
 import * as R from 'ramda';
 import type { AttributeDefinition } from '../../schema/attribute-definition';
 import { schemaAttributesDefinition } from '../../schema/schema-attributes';
-import { RELATION_ALLOWED_BY, RELATION_PARTICIPATE_TO } from '../../schema/internalRelationship';
+import { RELATION_ALLOWED_BY, RELATION_IN_PIR, RELATION_PARTICIPATE_TO } from '../../schema/internalRelationship';
 
 export const internalRelationshipsAttributes: { [k: string]: Array<AttributeDefinition> } = {
   [RELATION_PARTICIPATE_TO]: [
@@ -11,6 +11,32 @@ export const internalRelationshipsAttributes: { [k: string]: Array<AttributeDefi
   ],
   [RELATION_ALLOWED_BY]: [
     { name: 'grant', label: 'Grant', type: 'string', format: 'short', mandatoryType: 'no', editDefault: false, multiple: true, upsert: false, isFilterable: true }
+  ],
+  [RELATION_IN_PIR]: [
+    {
+      name: 'pir_score',
+      label: 'PIR Score',
+      type: 'numeric',
+      precision: 'integer',
+      mandatoryType: 'no',
+      editDefault: false,
+      multiple: false,
+      upsert: true,
+      isFilterable: true,
+      featureFlag: 'Pir',
+    },
+    {
+      name: 'pir_explanations',
+      label: 'PIR Explanations',
+      type: 'object',
+      format: 'flat',
+      mandatoryType: 'no',
+      editDefault: false,
+      multiple: true,
+      upsert: true,
+      isFilterable: true,
+      featureFlag: 'Pir',
+    },
   ],
 };
 R.forEachObjIndexed((value, key) => schemaAttributesDefinition.registerAttributes(key as string, value), internalRelationshipsAttributes);
