+++ opencti-platform/opencti-graphql/src/modules/relationsRef/stixCoreObject-registrationRef.ts
@@ -1,12 +1,8 @@
 import { schemaRelationsRefDefinition } from '../../schema/schema-relationsRef';
 import { ABSTRACT_STIX_CORE_OBJECT } from '../../schema/general';
-import { createdBy, externalReferences, inPir, objectLabel, objectMarking } from '../../schema/stixRefRelationship';
-import { isFeatureEnabled } from '../../config/conf';
+import { createdBy, externalReferences, objectLabel, objectMarking } from '../../schema/stixRefRelationship';
 
 const stixCoreObjectsMetaRel = [createdBy, objectMarking, objectLabel, externalReferences];
-if (isFeatureEnabled('Pir')) {
-  stixCoreObjectsMetaRel.push(inPir);
-}
 
 schemaRelationsRefDefinition.registerRelationsRef(
   ABSTRACT_STIX_CORE_OBJECT,
