+++ opencti-platform/opencti-graphql/src/schema/stixRefRelationship.ts
@@ -10,7 +10,6 @@ import {
   INPUT_ETHNICITY,
   INPUT_EXTERNAL_REFS,
   INPUT_GRANTED_REFS,
-  INPUT_IN_PIR,
   INPUT_INTERNAL_FILES,
   INPUT_KILLCHAIN,
   INPUT_LABELS,
@@ -53,9 +52,6 @@ import {
   ENTITY_WINDOWS_REGISTRY_VALUE_TYPE
 } from './stixCyberObservable';
 import { ATTRIBUTE_SAMPLE } from '../modules/malwareAnalysis/malwareAnalysis-types';
-import { ENTITY_TYPE_PIR, type StoreRelationPir } from '../modules/pir/pir-types';
-import { isFeatureEnabled } from '../config/conf';
-import type { StoreCommon } from '../types/store';
 
 export const ABSTRACT_STIX_NESTED_REF_RELATIONSHIP = 'stix-nested-ref-relationship'; // Only for front usage
 
@@ -664,7 +660,6 @@ export const RELATION_OBJECT_ASSIGNEE = 'object-assignee';
 export const RELATION_OBJECT_PARTICIPANT = 'object-participant';
 export const RELATION_BORN_IN = 'born-in'; // Extension (TIM)
 export const RELATION_ETHNICITY = 'of-ethnicity'; // Extension (TIM)
-export const RELATION_IN_PIR = 'in-pir';
 
 // EXTERNAL
 
@@ -906,24 +901,6 @@ export const killChainPhases: RefAttribute = {
   toTypes: [ENTITY_TYPE_KILL_CHAIN_PHASE],
 };
 
-export const inPir: RefAttribute = {
-  name: INPUT_IN_PIR,
-  type: 'ref',
-  databaseName: RELATION_IN_PIR,
-  stixName: 'in_pir_refs',
-  mandatoryType: 'no',
-  editDefault: false,
-  multiple: true,
-  upsert: false,
-  isRefExistingForTypes(this, _, toType) {
-    return this.toTypes.includes(toType);
-  },
-  label: 'In PIR',
-  datable: false,
-  isFilterable: false,
-  toTypes: [ENTITY_TYPE_PIR],
-};
-
 export const META_RELATIONS: RefAttribute[] = [
   objectLabel,
   externalReferences,
@@ -937,9 +914,6 @@ export const META_RELATIONS: RefAttribute[] = [
   objectOrganization,
   objectAssignee,
 ];
-if (isFeatureEnabled('Pir')) {
-  META_RELATIONS.push(inPir);
-}
 
 // Register
 schemaTypesDefinition.register(
@@ -948,10 +922,6 @@ schemaTypesDefinition.register(
 );
 
 export const isStixRefRelationship = (type: string) => schemaTypesDefinition.isTypeIncludedIn(type, ABSTRACT_STIX_REF_RELATIONSHIP) || type === ABSTRACT_STIX_REF_RELATIONSHIP;
-
-export const isStoreRelationPir = (instance: StoreCommon): instance is StoreRelationPir => {
-  return (instance as StoreRelationPir).entity_type === RELATION_IN_PIR;
-};
 
 export const buildRelationRef = (relationRef: Omit<RefAttribute, 'isRefExistingForTypes'>, isRefExistingForTypes: Checker): RefAttribute => {
   return {
